pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "meta-builder"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "deepagents",
    "anthropic[aiohttp]",
    "mcp[cli]",
    "redis[asyncio]",
    "pydantic>=2.0",
    "tavily-python",
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "mcp*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v"
markers = [
    "functional: real-API smoke tests, not mocked (deselect with -m 'not functional')",
]

[tool.ruff]
line-length = 100
target-version = "py312"
